requirements.txt
python-dotenv==1.0.1 
pandas==2.2.3 
streamlit==1.39.0 
requests==2.32.3 
llama-index==0.11.11 
jinja2==3.1.4 
fastapi==0.115.0 
pydantic==2.11.0
asyncpg==0.29.0
uvicorn==0.35.0
psutil==7.0.0
mcp==1.13.1
phidata==2.7.10
langchain-mcp-adapters==0.1.11
langgraph==0.6.8
psycopg-binary==3.2.10
langchain-openai==0.3.34
RestrictedPython==8.1
httpx==0.28.1